[fix] routes 변경
