+++ src/RootLayout.tsx
@@ -1,51 +1,13 @@
-import { Outlet, useLocation } from 'react-router';
-import Footer from './shared/components/footer/Footer';
-import Header from './shared/components/header/Header';
-import ScrollUpButton from './shared/components/button/ScrollUpButton';
-import clsx from 'clsx';
-import EasterEggSection from './features/easter-egg/components/EasterEggSection';
+import { Outlet } from 'react-router';
 import { AuthProvider } from './shared/utils/AuthProvider';
 import { ModalProvider } from './shared/utils/ModalProvider';
 
-// const TABS = [
-//   { id: 'all', label: '전체' },
-//   { id: 'play', label: '투표/게임' },
-// ];
 // 프로젝트 레이아웃
 const RootLayout = () => {
-  // const [tab, setTab] = useState('all');
-
-  const location = useLocation();
-  const isAdmin = location.pathname.startsWith('/admin');
-  const isThread = location.pathname.startsWith('/thread');
-
   return (
     <AuthProvider>
       <ModalProvider>
-        <div className="flex flex-col relative">
-          <Header
-            // tabs={isThread ? TABS : undefined}
-            // currentTab={tab}
-            // onTabChange={(tabId: string) => setTab(tabId)}
-            hideParticipantCount={isAdmin}
-          />
-
-          <div
-            className={clsx(
-              'pt-12 md:pt-15',
-              isThread ? 'root-min-h-thread' : 'root-min-h',
-            )}
-          >
-            <main className="relative bg-main">
-              <Outlet />
-            </main>
-          </div>
-
-          {!isThread && <Footer />}
-
-          <ScrollUpButton />
-          <EasterEggSection />
-        </div>
+        <Outlet />
       </ModalProvider>
     </AuthProvider>
   );
