+++ src/shared/styles/global.css
@@ -16,7 +16,7 @@
   font-display: swap;
 }
 
-/* HomeLayout 전용 클래스 */
+/* Home 전용 클래스 */
 .home-font {
   font-family: 'OngleipParkDahyeon', sans-serif;
 }
